Aadhar-Classification-main/Aadhar-Classification-main/requirements.txt
ultralytics>=8.3.0
huggingface_hub
supervision
opencv-python
pillow
matplotlib
numpy
requests
pytesseract
easyocr
ensemble-boxes
torch
python-dotenv
fastapi
uvicorn
python-multipart
hf_xet
PyMuPDF
imagehash
paddleocr
paddlepaddle
